More printing of the halmos version
Note: during CI builds, we only perform a sparse checkout so we expect the printed version to be slightly wonky, e.g. `halmos 0.1.dev1+g81b7ae5`

(base tag is wrong, number of commits ahead is wrong, but git hash is correct)

For released images, we expect the version string to be correct (e.g. `halmos 0.1.14`)