fix OZ ERC721 test
on the 5.x branch of openzeppelin-contracts, the authorization check in ERC721 transfers is optional, and no auth check is performed is auth == address(0). In transferFrom, auth is set to _msgSender(). That's why we kept generating counterexamples for these asserts:

   // ensure requirements of transfer
   assertEq(from, oldOwner);
   assertTrue(approved);

the counterexamples said that for caller == address(0), we didn't have to be approved and we could indeed transfer any tokens 🫠️️️️️️