avoid solver race conditions with a lock
ideally we would solve this after parsing the command line and before entering tests, but unfortunately this approach does not work because solvers can be overridden at the test level. So we could analyze the config when it's created at the different levels (command line, contract, function...) but this seems overly complex, and we may also run contracts and tests concurrently in the future so this may break.

Instead let's use a simple lock on `ensure_solver_available`, captured in the Java-style `synchronized` decorator. This time the flow is that `ensure_solver_available` can only be entered by one thread at a time, therefore:
- the first time we find a missing solver, we start the download
- other concurrent queries are waiting
- once the download is finished and the command is resolved, the lock is released
- the concurrent queries get the resolved command, without having to trigger the download again

fixes https://github.com/a16z/halmos/issues/568